Arn 1026 get risk management plans (#165)
* :wip: adding endpoint and dtos for risk management plan and timeline from ords (through offender-assessments)

* ARN-1038 Code to handle complete assessments in timeline but not in a… (#164)

* ARN-1038 Code to handle complete assessments in timeline but not in assessment block

* ARN-1038 Remove println from test

* rebase on main

* add happy path to get risk management plans

* add risk management service tests

* refactor risk management plan into risk controller

* fix typo

* refactor for readability

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/assessrisksandneeds/api/model/AssessmentOffenceDto.kt
@@ -19,7 +19,7 @@ data class AssessmentOffenceDto(
 data class TimelineDto(
 
   @JsonProperty("assessmentPk")
-  val assessmentId: Int,
+  val assessmentId: Long,
   val initiationDate: LocalDateTime,
   val status: String,
   val completedDate: LocalDateTime?,
@@ -30,7 +30,7 @@ data class TimelineDto(
 data class AssessmentDto(
 
   @JsonProperty("assessmentPk")
-  val assessmentId: Int,
+  val assessmentId: Long,
   val dateCompleted: LocalDateTime?,
   val initiationDate: LocalDateTime,
   val assessmentStatus: String,
